Sorted physics when shooting at weapons on floor
The amount that they moved is based upon the damage of the gun. More
powerful weapons cause more movement in the guns. I also removed sprint
for now as it just didn't feel right with its current implementation. I
also started putting the GUI elements in place. It doesn't work yet it
but it looks better and that is all I'm worried about until I get the
chance to start working on stuff properly.